pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mne-helper"
version = "0.1.0"
description = "EEG data processing and visualization tool"
readme = "README.md"
requires-python = ">=3.8"
keywords = ["eeg", "signal-processing", "machine-learning", "neuroscience"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = ["mne>=1.0.0", 
                "numpy>=1.20.0", 
                "python-dotenv>=0.19.0", 
                "scipy>=1.7.0", 
                "matplotlib>=3.5.0", 
                "scikit-learn>=1.0.0", 
                "pytest>=6.0.0", 
                "pytest-cov>=3.0.0", 
                "pre-commit>=2.15.0", 
                "ruff>=0.1.0"]


[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

